Add 'real-time' mode to generator
'Real-time' refers to the output frequency of the random prices, i.e. in
real-time mode it tries to adhere to the frequency given in the --freq
parameter, but still uses the provided starttime instead of using
wall-time for the timestamping.

Also add logging on provide way to influence verbosity of log output in
generator.